Fix dash enemy frame lag
did this by making a collider around the dash enemy that is a new layer and having it only collide with enemies. Made all enemies a part of Enemy layer which is same as default except it also collides with the DashEnemyExtra layer. Not 100% sure this works but in theory it should.