Created final boss portal tile
Tile has a chance to spawn in when you are a chosen amount of tiles away. A portal will be put in the tile that takes the player to the final boss room. Portal will have a black hole affect that pulls player to it. Maybe add a gate to the portal that requires you to have a certain amount of coins or score in the round to enter?