Continuous Data Between Scenes
Made a Scene Manager which switches between scenes and handles all of the resets needed in the Game Manager.  Made the Game Manager so it doesn't unload in between scenes in order to do this. Made a lot of changes in order to make this work which could result in bugs in the future.